Update batch_gen.py
fff

+++ axelerate/networks/yolo/backend/batch_gen.py
@@ -82,7 +82,7 @@ class BatchGenerator(Sequence):
                 annotation.append(labels[j])
                 annotations.append(annotation)
             anns_list.append(np.array(annotations))
-        return imgs_list, np.array(anns_list)
+        return imgs_list, np.array(anns_list,dtype=float32)
 
 
     def __getitem__(self, idx):
